Split the test file into separate ones.
