+++ tests/instruments/test_channel.py
@@ -1,0 +1,62 @@
+#
+# This file is part of the PyMeasure package.
+#
+# Copyright (c) 2013-2022 PyMeasure Developers
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+
+from unittest import mock
+
+import pytest
+
+from pymeasure.units import ureg
+from pymeasure.test import expected_protocol
+from pymeasure.instruments import Instrument, Channel
+from pymeasure.instruments.common_base import DynamicProperty, CommonBase
+from pymeasure.adapters import FakeAdapter, ProtocolAdapter
+from pymeasure.instruments.fakes import FakeInstrument
+from pymeasure.instruments.validators import strict_discrete_set, strict_range, truncated_range
+
+
+class TestChannelCommunication:
+    @pytest.fixture()
+    def ch(self):
+        a = mock.MagicMock(return_value="5")
+        return Channel(a, "A")
+
+    def test_write(self, ch):
+        ch.write("abc")
+        assert ch.parent.method_calls == [mock.call.write('abc')]
+
+    def test_read(self, ch):
+        ch.read()
+        assert ch.parent.method_calls == [mock.call.read()]
+
+    def test_write_bytes(self, ch):
+        ch.write_bytes(b"abc")
+        assert ch.parent.method_calls == [mock.call.write_bytes(b"abc")]
+
+    def test_read_bytes(self, ch):
+        ch.read_bytes(5)
+        assert ch.parent.method_calls == [mock.call.read_bytes(5)]
+
+    def test_write_binary_values(self, ch):
+        ch.write_binary_values("abc", [5, 6, 7])
+        assert ch.parent.method_calls == [mock.call.write_binary_values("abc", [5, 6, 7])]
